Fixed overflow issue that would only happen on 64 bit compiles with hash tables or geobuckets larger than 2^31. After this change, all remaining issues found by the static analysis of Visual Studio 2012 Professional are false positives.

+++ src/mathicgb/PolyHashTable.cpp
@@ -12,8 +12,8 @@ const bool AlwaysInsertAtEnd = true;
 
 PolyHashTable::PolyHashTable(const PolyRing *R, int nbits)
   : mRing(*R),
-    mHashMask((1 << nbits)-1),
-    mTableSize(1 << nbits),
+    mHashMask((static_cast<size_t>(1) << nbits)-1),
+    mTableSize(static_cast<size_t>(1) << nbits),
     mLogTableSize(nbits),
     mNodeCount(0),
     mBinCount(0),
